fix overflow

+++ dribas.sergey/S1/outSequance.hpp
@@ -7,8 +7,8 @@
 
 namespace dribas
 {
-  void getSequanceName(std::ostream&, const List< std::pair< std::string, List< int > > >&);
-  void getSequanceNameSum(std::ostream& out, const List< std::pair< std::string, List< int > > >&);
+  void getSequanceName(std::ostream&, const List< std::pair< std::string, List< unsigned long long > > >&);
+  void getSequanceNameSum(std::ostream& out, const List< std::pair< std::string, List< unsigned long long > > >&);
 }
 
 #endif
